More compiler warnings addressed (#5414).  ogr, 2nd part of fmts and apps
git-svn-id: https://svn.osgeo.org/gdal/branches/1.11@27741 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/apps/gdaltindex.c
@@ -95,7 +95,7 @@ int main(int argc, char *argv[])
     const char *index_filename = NULL;
     const char *tile_index = "location";
     const char* pszDriverName = "ESRI Shapefile";
-    int nMaxFieldSize = 254;
+    size_t nMaxFieldSize = 254;
     int		i_arg, ti_field;
     OGRDataSourceH hTileIndexDS;
     OGRLayerH hLayer = NULL;
